\name{addq}
\alias{addq}
\title{utility - adds a column (x[,1]-mean(x[,1]))**2}
\usage{
  addq(x)
}
\arguments{
  \item{x}{matrix}
}
\value{
  matrix with added column
}
\description{
  utility - adds a column (x[,1]-mean(x[,1]))**2
}
\section{Details}{

}
\author{
  Giles Heywood
}
\seealso{
  Other internals: \code{\link{buialigned}},
  \code{\link{pococe}}
}

