\name{buialigned}
\alias{buialigned}
\title{utility}
\usage{
  buialigned(x, ret)
}
\arguments{
  \item{x}{object of class ce}

  \item{ret}{return matrix}
}
\value{
  validity
}
\description{
  utility
}
\section{Details}{

}
\author{
  Giles Heywood
}
\seealso{
  Other internals: \code{\link{addq}}, \code{\link{pococe}}
}

