\name{buice}
\alias{buice}
\title{Extractor for identifiers (all)}
\usage{
  buice(x)
}
\arguments{
  \item{x}{object of class ce}
}
\value{
  identifiers : colnames from original data
}
\description{
  Extractor for identifiers (all)
}
\section{Details}{

}
\author{
  Giles Heywood
}
\seealso{
  Other extractors: \code{\link{devce}},
  \code{\link{face}}, \code{\link{fmpce}},
  \code{\link{fulce}}, \code{\link{genfrdce}},
  \code{\link{ldgce}}, \code{\link{ldggmace}},
  \code{\link{metce}}, \code{\link{mktce}},
  \code{\link{msce}}, \code{\link{msrtce}},
  \code{\link{prdce}}, \code{\link{quace}},
  \code{\link{resce}}, \code{\link{scoce}},
  \code{\link{sdvce}}, \code{\link{spvce}},
  \code{\link{stce}}, \code{\link{sysce}},
  \code{\link{unqce}}, \code{\link{vcvce}},
  \code{\link{vfuce}}
}

