\name{devce}
\alias{devce}
\title{Extractor for score on: deviation of rem from x-section mean}
\usage{
  devce(x, ret)
}
\arguments{
  \item{x}{object of class ce}

  \item{ret}{returns matrix}
}
\value{
  matrix
}
\description{
  Extractor for score on: deviation of rem from x-section
  mean
}
\section{Details}{

}
\author{
  Giles Heywood
}
\seealso{
  Other extractors: \code{\link{buice}},
  \code{\link{face}}, \code{\link{fmpce}},
  \code{\link{fulce}}, \code{\link{genfrdce}},
  \code{\link{ldgce}}, \code{\link{ldggmace}},
  \code{\link{metce}}, \code{\link{mktce}},
  \code{\link{msce}}, \code{\link{msrtce}},
  \code{\link{prdce}}, \code{\link{quace}},
  \code{\link{resce}}, \code{\link{scoce}},
  \code{\link{sdvce}}, \code{\link{spvce}},
  \code{\link{stce}}, \code{\link{sysce}},
  \code{\link{unqce}}, \code{\link{vcvce}},
  \code{\link{vfuce}}
}

