\name{face}
\alias{face}
\title{returns decomposed by all factors 1:k, + residual, + quadratic}
\usage{
  face(x, ret)
}
\arguments{
  \item{x}{object of class ce}

  \item{ret}{returns matrix}
}
\value{
  named list of attributed return: 1:k, ret, rer, qua
}
\description{
  returns decomposed by all factors 1:k, + residual, +
  quadratic
}
\section{Details}{

}
\author{
  Giles Heywood
}
\seealso{
  Other extractors: \code{\link{buice}},
  \code{\link{devce}}, \code{\link{fmpce}},
  \code{\link{fulce}}, \code{\link{genfrdce}},
  \code{\link{ldgce}}, \code{\link{ldggmace}},
  \code{\link{metce}}, \code{\link{mktce}},
  \code{\link{msce}}, \code{\link{msrtce}},
  \code{\link{prdce}}, \code{\link{quace}},
  \code{\link{resce}}, \code{\link{scoce}},
  \code{\link{sdvce}}, \code{\link{spvce}},
  \code{\link{stce}}, \code{\link{sysce}},
  \code{\link{unqce}}, \code{\link{vcvce}},
  \code{\link{vfuce}}
}

