% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce.R
\name{fmpce}
\alias{fmpce}
\title{Factor mimicking portfolio weights (but still in reduced dimension ie non-NA columns only)}
\usage{
fmpce(x, type = c("fmp", "hpl"))
}
\arguments{
\item{x}{object of class ce}

\item{type}{fmp for factor portfolio weights}
}
\value{
matrix
}
\description{
Factor mimicking portfolio weights (but still in reduced dimension ie non-NA columns only)
}
\section{Details}{

}
\author{
Giles Heywood
}
\seealso{
Other extractors: \code{\link{buice}}, \code{\link{devce}},
  \code{\link{face}}, \code{\link{fulce}},
  \code{\link{genfrdce}}, \code{\link{ldgce}},
  \code{\link{ldggmace}}, \code{\link{metce}},
  \code{\link{mktce}}, \code{\link{msce}},
  \code{\link{msrtce}}, \code{\link{prdce}},
  \code{\link{quace}}, \code{\link{resce}},
  \code{\link{scoce}}, \code{\link{sdvce}},
  \code{\link{spvce}}, \code{\link{stce}},
  \code{\link{sysce}}, \code{\link{unqce}},
  \code{\link{vcvce}}, \code{\link{vfuce}}
}

