% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce.R
\name{fulce}
\alias{fulce}
\title{Extractor for identifiers of columns with no NA (full data)}
\usage{
fulce(x)
}
\arguments{
\item{x}{object of class ce}
}
\value{
identifiers
}
\description{
Extractor for identifiers of columns with no NA (full data)
}
\section{Details}{

}
\author{
Giles Heywood
}
\seealso{
Other extractors: \code{\link{buice}}, \code{\link{devce}},
  \code{\link{face}}, \code{\link{fmpce}},
  \code{\link{genfrdce}}, \code{\link{ldgce}},
  \code{\link{ldggmace}}, \code{\link{metce}},
  \code{\link{mktce}}, \code{\link{msce}},
  \code{\link{msrtce}}, \code{\link{prdce}},
  \code{\link{quace}}, \code{\link{resce}},
  \code{\link{scoce}}, \code{\link{sdvce}},
  \code{\link{spvce}}, \code{\link{stce}},
  \code{\link{sysce}}, \code{\link{unqce}},
  \code{\link{vcvce}}, \code{\link{vfuce}}
}

