\name{genfrdce}
\alias{genfrdce}
\title{Factor risk decomposition by security}
\usage{
  genfrdce(x = getce())
}
\arguments{
  \item{x}{object of class ce}
}
\value{
  matrix, rows=security, columns = (total, residual,
  factors 1:k)
}
\description{
  Factor risk decomposition by security
}
\section{Details}{

}
\author{
  Giles Heywood
}
\seealso{
  Other extractors: \code{\link{buice}},
  \code{\link{devce}}, \code{\link{face}},
  \code{\link{fmpce}}, \code{\link{fulce}},
  \code{\link{ldgce}}, \code{\link{ldggmace}},
  \code{\link{metce}}, \code{\link{mktce}},
  \code{\link{msce}}, \code{\link{msrtce}},
  \code{\link{prdce}}, \code{\link{quace}},
  \code{\link{resce}}, \code{\link{scoce}},
  \code{\link{sdvce}}, \code{\link{spvce}},
  \code{\link{stce}}, \code{\link{sysce}},
  \code{\link{unqce}}, \code{\link{vcvce}},
  \code{\link{vfuce}}
}

