\name{ldggmace}
\alias{ldggmace}
\title{Extractor for loadings * fmp, (colnames=T, rownames=T)}
\usage{
  ldggmace(x)
}
\arguments{
  \item{x}{object of class ce}
}
\value{
  named list(M,S,R)
}
\description{
  Extractor for loadings * fmp, (colnames=T, rownames=T)
}
\section{Details}{

}
\author{
  Giles Heywood
}
\seealso{
  Other extractors: \code{\link{buice}},
  \code{\link{devce}}, \code{\link{face}},
  \code{\link{fmpce}}, \code{\link{fulce}},
  \code{\link{genfrdce}}, \code{\link{ldgce}},
  \code{\link{metce}}, \code{\link{mktce}},
  \code{\link{msce}}, \code{\link{msrtce}},
  \code{\link{prdce}}, \code{\link{quace}},
  \code{\link{resce}}, \code{\link{scoce}},
  \code{\link{sdvce}}, \code{\link{spvce}},
  \code{\link{stce}}, \code{\link{sysce}},
  \code{\link{unqce}}, \code{\link{vcvce}},
  \code{\link{vfuce}}
}

