% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce.R
\name{msrtce}
\alias{msrtce}
\title{Extractor for components Market, Systematic, Residual, Total}
\usage{
msrtce(x, ret)
}
\arguments{
\item{x}{object of class ce}

\item{ret}{returns matrix}
}
\value{
matrix of attributed return
}
\description{
Extractor for components Market, Systematic, Residual, Total
}
\section{Details}{
 the output has 4x columns of the input, grouped by component
}
\author{
Giles Heywood
}
\seealso{
Other extractors: \code{\link{buice}}, \code{\link{devce}},
  \code{\link{face}}, \code{\link{fmpce}},
  \code{\link{fulce}}, \code{\link{genfrdce}},
  \code{\link{ldgce}}, \code{\link{ldggmace}},
  \code{\link{metce}}, \code{\link{mktce}},
  \code{\link{msce}}, \code{\link{prdce}},
  \code{\link{quace}}, \code{\link{resce}},
  \code{\link{scoce}}, \code{\link{sdvce}},
  \code{\link{spvce}}, \code{\link{stce}},
  \code{\link{sysce}}, \code{\link{unqce}},
  \code{\link{vcvce}}, \code{\link{vfuce}}
}

