\name{pococe}
\alias{pococe}
\title{extractor for poco}
\usage{
  pococe(x)
}
\arguments{
  \item{x}{object of class ce}
}
\value{
  matrix
}
\description{
  extractor for poco
}
\section{Details}{

}
\author{
  Giles Heywood
}
\seealso{
  Other internals: \code{\link{addq}},
  \code{\link{buialigned}}
}

