\name{sdvce}
\alias{sdvce}
\title{Extractor for standard deviation}
\usage{
  sdvce(x)
}
\arguments{
  \item{x}{object of class ce}
}
\value{
  standard deviation (vol)
}
\description{
  Extractor for standard deviation
}
\section{Details}{

}
\author{
  Giles Heywood
}
\seealso{
  Other extractors: \code{\link{buice}},
  \code{\link{devce}}, \code{\link{face}},
  \code{\link{fmpce}}, \code{\link{fulce}},
  \code{\link{genfrdce}}, \code{\link{ldgce}},
  \code{\link{ldggmace}}, \code{\link{metce}},
  \code{\link{mktce}}, \code{\link{msce}},
  \code{\link{msrtce}}, \code{\link{prdce}},
  \code{\link{quace}}, \code{\link{resce}},
  \code{\link{scoce}}, \code{\link{spvce}},
  \code{\link{stce}}, \code{\link{sysce}},
  \code{\link{unqce}}, \code{\link{vcvce}},
  \code{\link{vfuce}}
}

