\name{spvce}
\alias{spvce}
\title{Extractor for pecific vol}
\usage{
  spvce(x)
}
\arguments{
  \item{x}{object of class ce}
}
\value{
  specific vol
}
\description{
  Extractor for pecific vol
}
\section{Details}{

}
\author{
  Giles Heywood
}
\seealso{
  Other extractors: \code{\link{buice}},
  \code{\link{devce}}, \code{\link{face}},
  \code{\link{fmpce}}, \code{\link{fulce}},
  \code{\link{genfrdce}}, \code{\link{ldgce}},
  \code{\link{ldggmace}}, \code{\link{metce}},
  \code{\link{mktce}}, \code{\link{msce}},
  \code{\link{msrtce}}, \code{\link{prdce}},
  \code{\link{quace}}, \code{\link{resce}},
  \code{\link{scoce}}, \code{\link{sdvce}},
  \code{\link{stce}}, \code{\link{sysce}},
  \code{\link{unqce}}, \code{\link{vcvce}},
  \code{\link{vfuce}}
}

