% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce.R
\name{stce}
\alias{stce}
\title{Market,Systematic,Residual,Total returns, divided by estimated vol}
\usage{
stce(x, ret)
}
\arguments{
\item{x}{object of class ce}

\item{ret}{returns matrix}
}
\value{
named list of attributed and normalised return
}
\description{
Market,Systematic,Residual,Total returns, divided by estimated vol
}
\section{Details}{
 components / vol estimate, in-sample
}
\author{
Giles Heywood
}
\seealso{
Other extractors: \code{\link{buice}}, \code{\link{devce}},
  \code{\link{face}}, \code{\link{fmpce}},
  \code{\link{fulce}}, \code{\link{genfrdce}},
  \code{\link{ldgce}}, \code{\link{ldggmace}},
  \code{\link{metce}}, \code{\link{mktce}},
  \code{\link{msce}}, \code{\link{msrtce}},
  \code{\link{prdce}}, \code{\link{quace}},
  \code{\link{resce}}, \code{\link{scoce}},
  \code{\link{sdvce}}, \code{\link{spvce}},
  \code{\link{sysce}}, \code{\link{unqce}},
  \code{\link{vcvce}}, \code{\link{vfuce}}
}

