% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce.R
\name{unqce}
\alias{unqce}
\title{Extractor for uniqueness}
\usage{
unqce(x)
}
\arguments{
\item{x}{object of class ce}
}
\description{
Extractor for uniqueness
}
\section{Details}{

}
\author{
Giles Heywood
}
\seealso{
Other extractors: \code{\link{buice}}, \code{\link{devce}},
  \code{\link{face}}, \code{\link{fmpce}},
  \code{\link{fulce}}, \code{\link{genfrdce}},
  \code{\link{ldgce}}, \code{\link{ldggmace}},
  \code{\link{metce}}, \code{\link{mktce}},
  \code{\link{msce}}, \code{\link{msrtce}},
  \code{\link{prdce}}, \code{\link{quace}},
  \code{\link{resce}}, \code{\link{scoce}},
  \code{\link{sdvce}}, \code{\link{spvce}},
  \code{\link{stce}}, \code{\link{sysce}},
  \code{\link{vcvce}}, \code{\link{vfuce}}
}

