\name{aace-package}
\alias{aace-package}
\alias{aace}
\docType{package}
\title{
Covariance Estimation
}
\description{
Covariance estimation using modified PCA, based on BurStFin
}
\details{
\tabular{ll}{
Package: \tab aace\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-04-16\cr
License: \GPL (>=2)\cr
}

}
\author{
Giles Heywood

Maintainer: gilesheywood at gmail.com
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
